fix(system2d.py): Fix Picart
Sign error caused opposite of Picart to be used.
